Re-design matmul interface for IVector
Redesgin the matmul interface for IVector to take weight and bias
as two input, because if we're using CoopVec in the future we will
have to keep those into two different storages because of the opaque
layout of the CoopVec matrix in the extension.

Though this makes the interface a little bit more complex, it can
support more general case.